`inline fun <reified T>` 함수 재정의 및 디렉토리 변경
