+++ Sunflower/app/src/main/java/com/jin/sunflower/core/data/RetrofitClient.kt
@@ -1,4 +1,4 @@
-package com.jin.sunflower.core.domain
+package com.jin.sunflower.core.data
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
@@ -8,12 +8,11 @@ import retrofit2.converter.gson.GsonConverterFactory
 object RetrofitClient {
     private val gson: Gson = GsonBuilder().create()
 
-    fun <T> createService(baseUrl: String, service: Class<T>): T {
-        val retrofitBuild = buildRetrofit(baseUrl)
-        return retrofitBuild.create(service)
+    inline fun <reified T> createService(baseUrl: String): T {
+        return buildRetrofit(baseUrl).create(T::class.java)
     }
 
-    private fun buildRetrofit(baseUrl: String): Retrofit {
+    fun buildRetrofit(baseUrl: String): Retrofit {
         return Retrofit.Builder()
             .baseUrl(baseUrl)
             .addConverterFactory(GsonConverterFactory.create(gson))
