+++ Sunflower/app/src/main/java/com/jin/sunflower/core/data/unsplash/UnsplashDataSource.kt
@@ -1,20 +1,20 @@
 package com.jin.sunflower.core.data.unsplash
 
 import com.jin.sunflower.BuildConfig
-import com.jin.sunflower.core.domain.unsplash.UnsplashService
+import com.jin.sunflower.core.domain.entity.UnsplashImage
 
 class UnsplashDataSource(private val apiService: UnsplashService) {
-    suspend fun searchPlantImageUrl(keyword: String): String {
+    suspend fun searchPlantImageUrl(keyword: String): UnsplashImage {
         return try {
             val response = apiService.fetchPlantImageUrl(
                 clientId = BuildConfig.UNSPLASH_API_KEY,
                 query = keyword,
                 orientation = "landscape"
             )
-            response.urls.regular
+            UnsplashImage(imageUrl = response.urls.regular)
         } catch (e: Exception) {
             println("Unsplash Error : ${e.message}")
-            "https://picsum.photos/300/200?random=1"
+            UnsplashImage(imageUrl = "https://picsum.photos/300/200?random=1")
         }
     }
 }
