+++ Sunflower/app/src/main/java/com/jin/sunflower/core/data/wikipedia/WikipediaResponse.kt
@@ -1,4 +1,4 @@
-package com.jin.sunflower.core.domain.wikipedia
+package com.jin.sunflower.core.data.wikipedia
 
 data class WikipediaResponse(
     val extract: String,
