+++ Sunflower/app/src/main/java/com/jin/sunflower/core/domain/entity/UnsplashImage.kt
@@ -1,0 +1,5 @@
+package com.jin.sunflower.core.domain.entity
+
+data class UnsplashImage(
+    val imageUrl: String
+)
