+++ Sunflower/app/src/main/java/com/jin/sunflower/core/domain/entity/WikipediaDescription.kt
@@ -1,0 +1,5 @@
+package com.jin.sunflower.core.domain.entity
+
+data class WikipediaDescription(
+    val description: String
+)
